Feat: find parties near 500m in party service

+++ src/party/party.entity.ts
@@ -27,13 +27,13 @@ export class Party {
   @Length(1, 100)
   restuarant: string;
 
-  @Column({ type: 'int' })
+  @Column({ type: 'decimal' })
   @IsNotEmpty()
   @IsNumber()
   @Min(0.0)
   meetLatitude: number;
 
-  @Column({ type: 'int' })
+  @Column({ type: 'decimal' })
   @IsNotEmpty()
   @IsNumber()
   @Min(0.0)
